@@ -178,7 +178,7 @@ impl CookieStore for Jar {
             .read()
             .unwrap()
             .get_request_values(url)
-            .map(|(name, value)| format!("{}={}", name, value))
+            .map(|(name, value)| format!("{name}={value}"))
             .collect::<Vec<_>>()
             .join("; ");
 
