@@ -52,7 +52,7 @@ fn new_resolver() -> io::Result<TokioAsyncResolver> {
     let (config, opts) = system_conf::read_system_conf().map_err(|e| {
         io::Error::new(
             io::ErrorKind::Other,
-            format!("error reading DNS system conf: {}", e),
+            format!("error reading DNS system conf: {e}"),
         )
     })?;
     Ok(TokioAsyncResolver::tokio(config, opts))
