@@ -12,9 +12,9 @@ where
     let mut buf = b"Basic ".to_vec();
     {
         let mut encoder = EncoderWriter::new(&mut buf, &BASE64_STANDARD);
-        let _ = write!(encoder, "{}:", username);
+        let _ = write!(encoder, "{username}:");
         if let Some(password) = password {
-            let _ = write!(encoder, "{}", password);
+            let _ = write!(encoder, "{password}");
         }
     }
     let mut header = HeaderValue::from_bytes(&buf).expect("base64 is always valid HeaderValue");
