@@ -12,7 +12,7 @@ async fn test_redirect_301_and_302_and_303_changes_post_to_get() {
     for &code in &codes {
         let redirect = server::http(move |req| async move {
             if req.method() == "POST" {
-                assert_eq!(req.uri(), &*format!("/{}", code));
+                assert_eq!(req.uri(), &*format!("/{code}"));
                 http::Response::builder()
                     .status(code)
                     .header("location", "/dst")
@@ -48,7 +48,7 @@ async fn test_redirect_307_and_308_tries_to_get_again() {
     for &code in &codes {
         let redirect = server::http(move |req| async move {
             assert_eq!(req.method(), "GET");
-            if req.uri() == &*format!("/{}", code) {
+            if req.uri() == &*format!("/{code}") {
                 http::Response::builder()
                     .status(code)
                     .header("location", "/dst")
@@ -90,7 +90,7 @@ async fn test_redirect_307_and_308_tries_to_post_again() {
             let data = req.body_mut().next().await.unwrap().unwrap();
             assert_eq!(&*data, b"Hello");
 
-            if req.uri() == &*format!("/{}", code) {
+            if req.uri() == &*format!("/{code}") {
                 http::Response::builder()
                     .status(code)
                     .header("location", "/dst")
@@ -127,7 +127,7 @@ fn test_redirect_307_does_not_try_if_reader_cannot_reset() {
     for &code in &codes {
         let redirect = server::http(move |mut req| async move {
             assert_eq!(req.method(), "POST");
-            assert_eq!(req.uri(), &*format!("/{}", code));
+            assert_eq!(req.uri(), &*format!("/{code}"));
             assert_eq!(req.headers()["transfer-encoding"], "chunked");
 
             let data = req.body_mut().next().await.unwrap().unwrap();
@@ -167,7 +167,7 @@ async fn test_redirect_removes_sensitive_headers() {
             let mid_addr = rx.borrow().unwrap();
             assert_eq!(
                 req.headers()["referer"],
-                format!("http://{}/sensitive", mid_addr)
+                format!("http://{mid_addr}/sensitive")
             );
             http::Response::default()
         }
@@ -179,7 +179,7 @@ async fn test_redirect_removes_sensitive_headers() {
         assert_eq!(req.headers()["cookie"], "foo=bar");
         http::Response::builder()
             .status(302)
-            .header("location", format!("http://{}/end", end_addr))
+            .header("location", format!("http://{end_addr}/end"))
             .body(Body::default())
             .unwrap()
     });
