@@ -20,5 +20,5 @@ async fn simple_example() {
     log(&format!("Status: {}", res.status()));
 
     let body = res.text().await.expect("response to utf-8 text");
-    log(&format!("Body:\n\n{}", body));
+    log(&format!("Body:\n\n{body}"));
 }
